Add feature to parse different time formats
We previously only allowed timers with the format of "hh:mm:ss" to be
used. With this patch, users can pass just seconds or minutes and
seconds.

Depending on how many colons are passed in the timer argument, we'll
figure out if the user is passing in just "ss", "mm:ss", or "hh:mm:ss".