Add window flash when timer is complete
If the countdown timer finishes while the program is not the active
window, the window will now flash to alert the user that the timer is
complete.

The "sdl2" dependency had to be updated to use the latest version on
github since "0.35.2" from crates.io did not include bindings for
SDL_FlashWindow.