Fix y-axis bounce in DVD Display mode
Since the font has some padding above it and below it, the bounce would
occur too early and the timer was never truly reaching the top or bottom
of the window. This fixes that by adding that padding into the bounce
calculation.

SDL TTF doesn't offer an easy way to find some of these values (such as
the distance from the descent to the bottom of the text), so the values
here are rough approximations -- but they're fine here for our purposes.