Create Desktop Notification for XDG (excluding macOS) versions
When the timer's finished, a Desktop Notification will be created to
alert the user. This will only occur on XDG (excluding macOS) versions
of the program. We exclude macOS because it has its own notification
system which doesn't work this way. But for all other 'unix' versions,
we can make the assumption that they follow the Desktop Notification
Specification (DNS) and we communicate with it using the D-Bus protocol.