Fix bug where countdown timer was too slow
On average, each second that was counted down from the timer took about
1.33 seconds of wall clock time. There were two reasons for this.

The first problem was sleeping at the wrong time. We would sleep before
preforming the time calcluation and we'd always sleep for the same
amount of time. The problem with this is that we weren't taking into
account the amount of time it actually took the CPU to perform the
calculation and render the timer to the screen. So on top of the time it
takes for the CPU to perform its calculations, we were always just
sleeping for a set amount of time. So even in the BEST case scenario
where the CPU execution time for a cycle was exactly 1/60th of a second,
we'd still sleep on top of it adding unnecessary time to the cycle.

The second problem is that we were sleeping at the wrong time. Since we
were sleeping before the heaviest part of a cycle (i.e. the rendering)
we wouldn't know how long the rendering took. The rendering time for a
cycle varies between cycles.

The solution here was to time the cycle and sleep for the remaining time
left in the 1/60th of a second. There is, however, a problem with this
solution. If the CPU is under a heavy load, then the cycle may take
longer than 1/60th of a second, and the timer will be off. A better
solution would be to not calculate the timer on the CPU, and instead,
read how much time has elapased during the cycle using using something
like `std::time`.

On average, this timer still isn't 100% wall clock accurate (each second
elapsed on the timer is about 1.03 seconds), but it's fine for my
current needs.