main: Fix timer inaccuracy
The timer would lag behind wall time ever so slightly (roughly 1 second
per minute). One problem with the original implementation was that the
timer was tracked separately from the frame time leading to a sleep at
the end of the frame in an attempt to catch up to wall time. The major
issue with this approach is that if the frame took longer than a 60th of
a second, then the timer would just fall behind wall time and it could
never catch up.

This approach throws out tracking FPS all together. The time it takes
for the frame to finish is calcluated (`delta_time`) and it's used to
increment the timer.